Add MetricsBehavior flag and allow setting no totalime
Currently the only way to make it so that no `totaltime` timeseries data for a metric is emitted is to set it at the `MetricsFactory` level with `TotaltimeType.NONE`.

However, many observability services (e.g. Lightstep, Datadog, Cloudwatch) charge per dimension and timeseries, so there may be cases where you want to more easily specify no `totaltime` for some metrics but not for others in order to reduce usage.

This adds a new `MetricsBehavior` enum that can be specified per metric.  For now, the only options for this are `DEFAULT` and `NO_TOTALTIME` (which, as the name suggests, means it won't include `totaltime`). In the future we may want to add an option e.g. to not include "prescient"/"environment" dimensions, among other possibilities.