feat: add support for Otlp exponential histograms (#38)
Adds support for Otlp exponential histograms. This is more or less the JVM implementation
of the goodmetrics_rs feature.

Unit tests were grabbed from the latest main of the Rust implementation's unit tests.

I also added a new sealed interface called DistributionMode so consumers can explicitly set which aggregation mode they would
like to use. Considering exponential histograms are a "newer" technology, the default is set to Histogram so this new version
is not a breaking change.